requirements.txt
Flask==2.3.3
netmiko==4.2.0
paramiko==3.3.1
flask-cors==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
